CIRCLE-33523 Fix finding the base revision on the first commit
To run a git diff, we need to give Git two refs to compare between. If
the current commit is the very first commit in the repo, its ref is
HEAD, but HEAD~1 (or ^1) does not exist.

In this case there is this magic SHA that Git uses for the empty tree
without any commits that we use as base revision instead.